+++ frontend/src/pages/about/About.tsx
@@ -2,7 +2,7 @@ import { Alert, Button, Link, Divider, SvgIconTypeMap, SxProps, Typography } fro
 import PageHeader from '../../components/PageHeader';
 import PageSection from '../../components/PageSection';
 import Logo from '../../components/logo/Logo';
-import CardflowTabs from '../../components/sellListing/CardflowTabs';
+import CardflowTabs from '../../components/cardflowTabs/CardflowTabs';
 import { theme } from '../../constants/theme';
 import ArrowRightIcon from '@mui/icons-material/ArrowRight';
 import FavoriteIcon from '@mui/icons-material/Favorite';
